Customers and payments model, schema, controller, endpoint update
I created the customers and payments table with all the needed (as of now) stuff for API endpoints. Some things don't work yet because we are done. Order ids needs to connect to payment for example but orders are not finished yet. I also added the relationship between orders and payments/customers